Refactor location area logic into LocationService
Removed ILocationAreaService and LocationAreaService, moving area calculation methods into ILocationService and LocationService. Updated InitiativeService to use ILocationService for area calculations. Cleaned up related dependency registrations and interface usages. Also moved GeneralStatisticsService to a new Statistics namespace and updated .editorconfig with unused code detection rules.
