+++ src/Application/Interfaces/Services/IGeneralStatisticsService.cs
@@ -3,7 +3,6 @@
 using System.Threading;
 using System.Threading.Tasks;
 
-using IAVH.BioTablero.CM.Application.DTOs.Reports;
 using IAVH.BioTablero.CM.Application.Utils;
 
 /// <summary>
