+++ src/WebApi/Config/DependencyRegistry/ConfigAppServices.cs
@@ -1,10 +1,10 @@
 ﻿namespace IAVH.BioTablero.CM.WebApi.Config.DependencyRegistry;
 
 using IAVH.BioTablero.CM.Application.Interfaces.Services;
-using IAVH.BioTablero.CM.Application.Services;
 using IAVH.BioTablero.CM.Application.Services.Geo;
 using IAVH.BioTablero.CM.Application.Services.Initiatives;
 using IAVH.BioTablero.CM.Application.Services.Logging;
+using IAVH.BioTablero.CM.Application.Services.Statistics;
 
 using Microsoft.Extensions.DependencyInjection;
 
@@ -31,7 +31,6 @@ public static class ConfigAppServices
         services.AddScoped<IJoinRequestService, JoinRequestService>();
         services.AddScoped<IJoinInvitationService, JoinInvitationService>();
         services.AddScoped<IGeneralStatisticsService, GeneralStatisticsService>();
-        services.AddScoped<ILocationAreaService, LocationAreaService>();
 
         return services;
     }
